+++ src/main/kotlin/com/yourssu/scouter/hrms/application/domain/member/MemberSyncResponse.kt
@@ -1,0 +1,5 @@
+package com.yourssu.scouter.hrms.application.domain.member
+
+data class MemberSyncResponse(
+    val failureMessages: List<String>,
+)
